+++ src/main/java/org/cheonyakplanet/be/application/dto/subscriprtion/CoordinateResponseDTO.java
@@ -1,14 +1,15 @@
-package org.cheonyakplanet.be.application.dto;
+package org.cheonyakplanet.be.application.dto.subscriprtion;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
-@Getter @Setter
+@Getter
+@Setter
 @RequiredArgsConstructor
 @AllArgsConstructor
 public class CoordinateResponseDTO {
-    private String x;
-    private String y;
+	private String x;
+	private String y;
 }
