+++ src/main/java/org/cheonyakplanet/be/application/service/InfoService.java
@@ -19,12 +19,12 @@ import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionDTO;
 import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionDetailDTO;
 import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionInfoSimpleDTO;
 import org.cheonyakplanet.be.application.dto.subscriprtion.SubscriptionLikeDTO;
-import org.cheonyakplanet.be.domain.entity.PublicFacility;
-import org.cheonyakplanet.be.domain.entity.School;
-import org.cheonyakplanet.be.domain.entity.Station;
-import org.cheonyakplanet.be.domain.entity.SubscriptionInfo;
-import org.cheonyakplanet.be.domain.entity.SubscriptionLike;
-import org.cheonyakplanet.be.domain.entity.SubscriptionLocationInfo;
+import org.cheonyakplanet.be.domain.entity.infra.PublicFacility;
+import org.cheonyakplanet.be.domain.entity.infra.School;
+import org.cheonyakplanet.be.domain.entity.infra.Station;
+import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionInfo;
+import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionLike;
+import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionLocationInfo;
 import org.cheonyakplanet.be.domain.entity.user.User;
 import org.cheonyakplanet.be.domain.repository.PublicFacilityRepository;
 import org.cheonyakplanet.be.domain.repository.SchoolRepository;
