+++ src/main/java/org/cheonyakplanet/be/domain/entity/infra/School.java
@@ -1,4 +1,4 @@
-package org.cheonyakplanet.be.domain.entity;
+package org.cheonyakplanet.be.domain.entity.infra;
 
 import org.cheonyakplanet.be.domain.Stamped;
