+++ src/main/java/org/cheonyakplanet/be/domain/entity/subscription/SubscriptionLike.java
@@ -1,4 +1,4 @@
-package org.cheonyakplanet.be.domain.entity;
+package org.cheonyakplanet.be.domain.entity.subscription;
 
 import java.time.LocalDate;
