+++ src/main/java/org/cheonyakplanet/be/domain/repository/HouseLoanRepository.java
@@ -1,6 +1,6 @@
 package org.cheonyakplanet.be.domain.repository;
 
-import org.cheonyakplanet.be.domain.entity.HouseLoan;
+import org.cheonyakplanet.be.domain.entity.finance.HouseLoan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
