+++ src/main/java/org/cheonyakplanet/be/domain/repository/MortgageRepository.java
@@ -1,6 +1,6 @@
 package org.cheonyakplanet.be.domain.repository;
 
-import org.cheonyakplanet.be.domain.entity.Mortgage;
+import org.cheonyakplanet.be.domain.entity.finance.Mortgage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
