+++ src/main/java/org/cheonyakplanet/be/domain/repository/PublicFacilityRepository.java
@@ -2,7 +2,7 @@ package org.cheonyakplanet.be.domain.repository;
 
 import java.util.List;
 
-import org.cheonyakplanet.be.domain.entity.PublicFacility;
+import org.cheonyakplanet.be.domain.entity.infra.PublicFacility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PublicFacilityRepository extends JpaRepository<PublicFacility, Long> {
