+++ src/main/java/org/cheonyakplanet/be/domain/repository/StationRepository.java
@@ -2,7 +2,7 @@ package org.cheonyakplanet.be.domain.repository;
 
 import java.util.List;
 
-import org.cheonyakplanet.be.domain.entity.Station;
+import org.cheonyakplanet.be.domain.entity.infra.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
