+++ src/main/java/org/cheonyakplanet/be/domain/repository/SubscriptionLikeRepository.java
@@ -3,7 +3,7 @@ package org.cheonyakplanet.be.domain.repository;
 import java.time.LocalDate;
 import java.util.List;
 
-import org.cheonyakplanet.be.domain.entity.SubscriptionLike;
+import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SubscriptionLikeRepository extends JpaRepository<SubscriptionLike, Long> {
