+++ src/main/java/org/cheonyakplanet/be/domain/repository/SubscriptionLocationInfoRepository.java
@@ -1,6 +1,6 @@
 package org.cheonyakplanet.be.domain.repository;
 
-import org.cheonyakplanet.be.domain.entity.SubscriptionLocationInfo;
+import org.cheonyakplanet.be.domain.entity.subscription.SubscriptionLocationInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SubscriptionLocationInfoRepository extends JpaRepository<SubscriptionLocationInfo, Long> {
