+++ src/main/java/org/cheonyakplanet/be/presentation/controller/UserController.java
@@ -8,6 +8,7 @@ import org.cheonyakplanet.be.application.dto.user.InterestLocationDTO;
 import org.cheonyakplanet.be.application.dto.user.LoginRequestDTO;
 import org.cheonyakplanet.be.application.dto.user.MyPageDTO;
 import org.cheonyakplanet.be.application.dto.user.SignupRequestDTO;
+import org.cheonyakplanet.be.application.dto.user.TokenResponse;
 import org.cheonyakplanet.be.application.dto.user.UserDTO;
 import org.cheonyakplanet.be.application.dto.user.UserUpdateRequestDTO;
 import org.cheonyakplanet.be.application.service.UserService;
@@ -82,16 +83,17 @@ public class UserController {
 		return ResponseEntity.ok(new ApiResponse("success", result));
 	}
 
-	@GetMapping("/kako/callback")
+	@GetMapping("/kakao/callback")
 	@Operation(summary = "소셜 로그인 - 카카오", description = "미완성")
-	public void kakaoLogin(@RequestParam String code, HttpServletResponse response) throws IOException {
-		String email = userService.kakaoLogin(code);
-
-		// 2. 저장된 Access Token 가져오기
-		String accessToken = jwtUtil.getAccessToken(email);
-
-		// 3. 프론트엔드로 리다이렉트 (Access Token 포함)
-		response.sendRedirect("https://frontend-domain.com/oauth/callback?accessToken=" + accessToken);
+	public void kakaoLogin(@RequestParam("code") String code, HttpServletResponse response) throws IOException {
+		TokenResponse tokens = userService.kakaoLogin(code);
+		String redirectUrl = String.format(
+			"https://cheongyakplanet.vercel.app?accessToken=%s&refreshToken=%s",
+			tokens.getAccessToken(),
+			tokens.getRefreshToken()
+		);
+		response.sendRedirect(redirectUrl);
+		log.error("토큰 검사 {}, {}", tokens.getAccessToken(), tokens.getRefreshToken());
 
 	}
