+++ src/main/java/org/cheonyakplanet/be/presentation/controller/kakaoController.java
@@ -1,0 +1,13 @@
+package org.cheonyakplanet.be.presentation.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class kakaoController {
+
+	@GetMapping("loginpage")
+	public String loginpage() {
+		return "login";
+	}
+}
