+++ src/main/resources/static/js/basic.js
@@ -1,6 +1,6 @@
 $(document).ready(function () {
     const auth = getToken();
-    if(auth === '') {
+    if (auth === '') {
         $('#login-true').hide();
         $('#login-false').show();
     } else {
@@ -12,16 +12,24 @@ $(document).ready(function () {
 let host = 'http://' + window.location.host;
 
 function logout() {
-    // 토큰 삭제
-    Cookies.remove('Authorization', { path: '/' });
-    window.location.href = host + "/api/member/login-page";
+    //     토큰    삭제
+
+    Cookies.remove('Authorization', {path: '/'});
+    window.location.href = host + '/api/user/login-page';
 }
 
 function getToken() {
-    let auth = Cookies.get('auth');
 
-    if(auth === undefined) {
+    let auth = Cookies.get('Authorization');
+
+    if (auth === undefined) {
         return '';
+    }
+
+    // kakao     로그인    사용한    경우 Bearer    추가
+
+    if (auth.indexOf('Bearer') === -1 && auth !== '') {
+        auth = 'Bearer ' + auth;
     }
 
     return auth;
